+++ CHANGELOG.rst
@@ -3,3 +3,8 @@
 Changelog
 =========
 TODO
+
+
+*   Test mutual exclusion
+
+*   Test & fix use of buffer